third_party/WebKit/LayoutTests/platform/win7/external/wpt/web-animations/timing-model/timelines/document-timelines-expected.txt

This is a testharness.js-based test.
FAIL Document timelines report current time relative to navigationStart assert_equals: The current time matches requestAnimationFrame time expected 15.051999999741383 but got 23.856666665778903
Harness: the test ran to completion.
